requirements.txt
wget
PyYAML
numpy
pillow
pytest
PyQt5
vispy
matplotlib
opencv-python
nuscenes-devkit